Making sun and terrain use materials. Adding IsLit property to Material. Sun now renders correctly.

+++ Source/ProspectEngine/Include/Material.h
@@ -12,7 +12,7 @@ namespace Prospect
       Material(std::shared_ptr<Material_impl>& impl);
 
       const Color& GetDiffuse() const;
-      unsigned long GetID() const;
+      unsigned GetID() const;
 
    private:
       std::shared_ptr<Material_impl> m_impl;
